Added basic settings for pupil algorithms, like initial image resize interpolation and run size. Upgraded Pylon dependency version to 8.1.0. Upgraded Qt version to 6.8.3 LTS. Upgraded qcustomplot. Trying out QJsonModel to store setup info. Slight changes on setup input GUI, still under construction.

+++ src/pupil-detection-methods/Pupil.h
@@ -120,23 +120,23 @@ public:
         return size.width > 0 && size.height > 0;
     }
 
-    int width() const {
-        return (int)size.width;
+    float width() const {
+        return size.width;
     }
 
-    int height() const {
-        return (int)size.height;
+    float height() const {
+        return size.height;
     }
 
-    int majorAxis() const {
-        return std::max<int>(size.width, size.height);
+    float majorAxis() const {
+        return std::max<float>(size.width, size.height);
     }
 
-    int minorAxis() const {
-        return std::min<int>(size.width, size.height);
+    float minorAxis() const {
+        return std::min<float>(size.width, size.height);
     }
 
-    int diameter() const {
+    float diameter() const {
         return majorAxis();
     }
